+++ architecture-project/db/demo.lock.db
@@ -1,4 +1,4 @@
 #FileLock
-#Wed Feb 24 09:27:39 MSK 2021
+#Wed Feb 24 19:14:00 MSK 2021
 method=file
-id=177d2b8baadbf941b6ef720b36611a2cac66a0d0bf3
+id=177d4d193a32b02022a73f9cdcd79f2d936c25ac021
