+++ architecture-project/src/main/java/ru/geekbrains/utils/ProductPriceObserver.java
@@ -1,13 +1,12 @@
 package ru.geekbrains.utils;
 
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
+import lombok.extern.java.Log;
 
+@Log
 public class ProductPriceObserver implements Observer{
 
-    private static final Logger logger = LoggerFactory.getLogger(ProductPriceObserver.class);
     @Override
     public void update(Subject subject, Object arg) {
-        logger.info("Price " + arg + "\n");
+        log.info("Price " + arg + "\n");
     }
 }
