+++ architecture-project/src/main/resources/templates/cart.html
@@ -18,8 +18,8 @@
     <div class="row">
         <div class="col-lg-12">
             <h1 class="my-4">Корзина</h1>
-            <span th:if="${@cartService.orderEntries.isEmpty()}">Ваша корзина пуста!</span>
-            <table th:if="${!@cartService.orderEntries.isEmpty()}" class="table table-hover">
+            <span th:if="${@cartServiceRealization.orderEntries.isEmpty()}">Ваша корзина пуста!</span>
+            <table th:if="${!@cartServiceRealization.orderEntries.isEmpty()}" class="table table-hover">
                 <thead class="thead-dark">
                 <tr>
                     <th>Название</th>
@@ -28,7 +28,7 @@
                 </tr>
                 </thead>
                 <tbody>
-                <tr th:each="orderEntry : ${@cartService.orderEntries}">
+                <tr th:each="orderEntry : ${@cartServiceRealization.orderEntries}">
                     <td th:text="${orderEntry.product.title}"/>
                     <td>
                         <a class="btn btn-danger" th:href="@{'/cart/removeOne/' + ${orderEntry.product.id}}">-</a>
@@ -41,11 +41,11 @@
                 <tr>
                     <td></td>
                     <td>Итого:</td>
-                    <td th:text="${@cartService.totalPrice}"></td>
+                    <td th:text="${@cartServiceRealization.totalPrice}"></td>
                 </tr>
                 </tbody>
             </table>
-            <div th:if="${!@cartService.orderEntries.isEmpty()}">
+            <div th:if="${!@cartServiceRealization.orderEntries.isEmpty()}">
                 <a class="btn btn-primary" th:href="@{/cart/createOrder}" th:text="#{text.cart.create}" sec:authorize="isAuthenticated()"></a>
                 <a class="btn btn-danger" th:href="@{/cart/clearCart}" th:text="#{text.cart.clear}"></a>
             </div>
