+++ architecture-project/src/main/resources/templates/fragments/navigation.html
@@ -16,7 +16,7 @@
                     <a class="nav-link" th:href="@{/admin/products}">Продукты</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" th:href="@{/cart}" th:text="#{text.cart.total(${@cartService.totalQuantity}, ${@cartService.totalPrice})}"/>
+                    <a class="nav-link" th:href="@{/cart}" th:text="#{text.cart.total(${@cartServiceRealization.totalQuantity}, ${@cartServiceRealization.totalPrice})}"/>
                 </li>
                 <li class="nav-item" th:classappend="${activeTab == 'Orders' ? 'active' : ''}">
                     <a class="nav-link" th:href="@{/admin/orders}">Заказы</a>
@@ -55,7 +55,7 @@
                         </a>
                     </li>
                     <li class="nav-item">
-                        <a class="nav-link" th:href="@{/cart}" th:text="#{text.cart.total(${@cartService.totalQuantity}, ${@cartService.totalPrice})}"/>
+                        <a class="nav-link" th:href="@{/cart}" th:text="#{text.cart.total(${@cartServiceRealization.totalQuantity}, ${@cartServiceRealization.totalPrice})}"/>
                     </li>
                     <li class="nav-item" sec:authorize="isAuthenticated()">
                         <a class="nav-link" sec:authentication="principal.username"></a>
